Store N, Z, and C flags as a copy of the AVR status register
Previously each was stored in a separate register, thus saving 2 regs.

Effect on cycle counts:
  -3: SBC ASL LSR (SLO SRE ASR)
  -2: ADC CMP CPX CPY ROL ROR (RLA RRA DCP ISB SBX)
  -1: (ANC)
  +1: LDA LDX LDY TXA TAX TYA TAY TSX TXS PLA BRK PHP (LAX)
  +3: BIT

ADC-imm was previously a bottleneck; this is a significant improvement.

Load and transfer instructions are a bit slower, but they are still
plenty fast enough (4 cycles for 2-tick instrs). BIT should also still
be fast enough; it only comes in zpg (3-tick) and abs (4-tick) versions.
We could inline the addressing code if it is a problem.

Decimal mode for ADC is temporarily disabled.